backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
scikit-learn==1.3.2
gunicorn==21.2.0
